+++ internal/domain/summary.go
@@ -5,19 +5,19 @@ import "time"
 // OverallBalance is the authenticated user's aggregated balance across all of
 // their groups. Net is owed minus owe (positive means they are owed money).
 type OverallBalance struct {
-	Net  float64 `json:"net"`
-	Owed float64 `json:"owed"`
-	Owe  float64 `json:"owe"`
+	Net  int64 `json:"net"`  // cents
+	Owed int64 `json:"owed"` // cents
+	Owe  int64 `json:"owe"`  // cents
 }
 
 // GroupSummary is a group as shown on the home screen, with the current user's
 // net balance in it.
 type GroupSummary struct {
-	ID           uint    `json:"id"`
-	Name         string  `json:"name"`
-	Emoji        string  `json:"emoji"`
-	MembersCount int     `json:"members_count"`
-	YourBalance  float64 `json:"your_balance"`
+	ID           uint   `json:"id"`
+	Name         string `json:"name"`
+	Emoji        string `json:"emoji"`
+	MembersCount int    `json:"members_count"`
+	YourBalance  int64  `json:"your_balance"` // cents
 }
 
 // HomeSummary powers the home screen in a single request.
@@ -35,7 +35,7 @@ type ActivityEvent struct {
 	Title      string    `json:"title"`
 	ActorID    uint      `json:"actor_id"`
 	ActorName  string    `json:"actor_name"`
-	Amount     float64   `json:"amount"`
-	YourShare  float64   `json:"your_share"`
+	Amount     int64     `json:"amount"`     // cents
+	YourShare  int64     `json:"your_share"` // cents
 	Date       time.Time `json:"date"`
 }
